Header cleanup
